Move settings call to dataset service, remove generation of stats
