+++ src/main/java/bio/terra/service/filedata/azure/AzureSynapseService.java
@@ -1,10 +1,5 @@
 package bio.terra.service.filedata.azure;
 
-import static bio.terra.service.filedata.azure.AzureSynapsePdao.getDataSourceName;
-
-import bio.terra.common.iam.AuthenticatedUserRequest;
-import bio.terra.model.AccessInfoModel;
-import bio.terra.service.dataset.Dataset;
 import bio.terra.service.resourcemanagement.MetadataDataAccessUtils;
 import org.springframework.stereotype.Component;
 
@@ -17,30 +12,5 @@ public class AzureSynapseService {
       AzureSynapsePdao azureSynapsePdao, MetadataDataAccessUtils metadataDataAccessUtils) {
     this.azureSynapsePdao = azureSynapsePdao;
     this.metadataDataAccessUtils = metadataDataAccessUtils;
-  }
-
-  /**
-   * @param dataset the dataset to configure the AzureDataSourceFor
-   * @param userRequest the user making the request
-   * @return the name of the datasource created
-   * @throws RuntimeException when the external datasource could not be configured
-   */
-  public String getOrCreateExternalAzureDataSource(
-      Dataset dataset, AuthenticatedUserRequest userRequest) {
-    String datasourceName = getDataSourceName(dataset.getId(), userRequest.getEmail());
-    AccessInfoModel accessInfoModel =
-        metadataDataAccessUtils.accessInfoFromDataset(dataset, userRequest);
-    String credName = AzureSynapsePdao.getCredentialName(dataset.getId(), userRequest.getEmail());
-
-    String metadataUrl =
-        "%s?%s"
-            .formatted(
-                accessInfoModel.getParquet().getUrl(), accessInfoModel.getParquet().getSasToken());
-    try {
-      azureSynapsePdao.getOrCreateExternalDataSource(metadataUrl, credName, datasourceName);
-      return datasourceName;
-    } catch (Exception e) {
-      throw new RuntimeException("Could not configure external datasource", e);
-    }
   }
 }
